requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# HTTP Client
httpx==0.25.2

# Data Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging and Utilities
python-multipart==0.0.6

# Optional: Redis (if you plan to use caching)
redis==5.0.1

# Optional: Development tools
black==23.11.0
isort==5.12.0